Moved common macro to a separate place
