+++ src/prodbg/viewdock/src/dock/serialize.rs
@@ -2,45 +2,6 @@ extern crate serde;
 use super::{Dock, DockHandle};
 use rect::Rect;
 
-macro_rules! gen_handle {
-    ($name:expr, $type_name:ident, $visitor:ident) => {
-        impl serde::ser::Serialize for $type_name {
-            fn serialize<S>(&self, serializer: &mut S) -> Result<(), S::Error>
-                where S: serde::ser::Serializer {
-                    serializer.serialize_newtype_struct($name, &self.0)
-                }
-        }
-
-        struct $visitor;
-
-        impl serde::Deserialize for $type_name  {
-            fn deserialize<D>(deserializer: &mut D) -> Result<$type_name, D::Error> where D: serde::de::Deserializer {
-                deserializer.deserialize_newtype_struct($name, $visitor)
-            }
-        }
-
-        impl serde::de::Visitor for $visitor {
-            type Value = $type_name;
-
-            fn visit_newtype_struct<D>(&mut self, deserializer: &mut D) -> Result<Self::Value, D::Error>
-                where D: serde::de::Deserializer {
-                    let value = try!(serde::de::Deserialize::deserialize(deserializer));
-                    Ok($type_name(value))
-                }
-
-            fn visit_seq<V>(&mut self, mut visitor: V) -> Result<$type_name, V::Error>
-                where V: serde::de::SeqVisitor {
-                    let v = match try!(visitor.visit()) {
-                        Some(value) => { value }
-                        None => { return Err(serde::de::Error::end_of_stream()); }
-                    };
-                    try!(visitor.end());
-                    Ok($type_name(v))
-                }
-        }
-    }
-}
-
 gen_handle!("DockHandle", DockHandle, DockHandleVisitor);
 
 // Serialization
